requirements.txt
Django
django-unfold==0.19.0
environs==10.0.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==24.2
python-dateutil~=2.9.0.post0
drf-standardized-errors==0.13.0
drf-yasg==1.21.7
gunicorn
django-cors-headers==4.3.1
pillow==10.3.0
